GravityMaster
Main window. 
Conditional statement for checking student's answer needs to be changed for minute degrees of error (off by 1, etc.) Might add significant figures to decrease the chances for error. 
(Note: current fonts in window incompatible with Eclipse Luna. When the executable jar is made, it must be exported with Eclipse Neon.)